corrigindo erro no arquivo desafio16.py

+++ desafios/desafio16.py
@@ -1,4 +1,4 @@
 import math
-num = int(input("Digite um número: "))
+num = float(input("Digite um número: "))
 inteiro = math.floor(num)
 print(f"O número digitado foi {num} e a sua porção inteira é {inteiro}")
